requirements.txt
Django==1.11.2
django-crispy-forms==1.6.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-taggit==0.22.1
markdown2==2.3.4
mysqlclient==1.3.10
pytz==2017.2
